Move config files back to frontend root
