+++ frontend/locales.ts
